+++ anxwritter/errors.py
@@ -89,6 +89,11 @@ class ErrorType(str, Enum):
     INVALID_CATEGORICAL_STYLE = 'invalid_categorical_style'
     STYLING_CONFLICT = 'styling_conflict'
 
+    # Canvas display (ANB v9 AttTime canvas-render workaround) errors
+    ATTTIME_VISIBLE_FORBIDS_CANVAS_DISPLAY = 'atttime_visible_forbids_canvas_display'
+    CANVAS_DISPLAY_INVALID = 'canvas_display_invalid'
+    CANVAS_DISPLAY_NAME_COLLISION = 'canvas_display_name_collision'
+
 
 class ANXValidationError(Exception):
     """Raised when chart data contains validation errors.